+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 ()
 
 - Add validation for `id` field of PathElement.
+- Add validation for `path` field of Key.
 
 
 ## 0.1.5 (2023-10-26)